+++ src/variables.js
@@ -1,8 +1,6 @@
-var dhlogo = null;
 var dhbanner = null;
 var dhvtcbg = null;
 var dhcustombg = (localStorage.getItem("custom-background") !== null ? localStorage.getItem("custom-background") : "");
-exports.dhlogo = dhlogo;
 exports.dhbanner = dhbanner;
 exports.dhvtcbg = dhvtcbg;
 exports.dhcustombg = dhcustombg;